change default CRD_URL to CRD_EXPERIMENTAL_URL
