+++ test/consts/zz_generated_gateway.go
@@ -3,7 +3,7 @@
 package consts
 
 const (
-	GatewayStandardCRDsKustomizeURL = "github.com/kubernetes-sigs/gateway-api/config/crd/?ref=v0.5.0"
-	GatewayCRDsKustomizeURL         = "github.com/kubernetes-sigs/gateway-api/config/crd/experimental?ref=v0.5.0"
-	GatewayRawRepoURL               = "https://raw.githubusercontent.com/kubernetes-sigs/gateway-api/v0.5.0"
+	GatewayStandardCRDsKustomizeURL     = "github.com/kubernetes-sigs/gateway-api/config/crd/?ref=v0.5.0"
+	GatewayExperimentalCRDsKustomizeURL = "github.com/kubernetes-sigs/gateway-api/config/crd/experimental?ref=v0.5.0"
+	GatewayRawRepoURL                   = "https://raw.githubusercontent.com/kubernetes-sigs/gateway-api/v0.5.0"
 )
