+++ test/internal/cmd/generate-gateway-api-urls/gateway_consts.tmpl
@@ -3,7 +3,7 @@
 package consts
 
 const (
-	GatewayStandardCRDsKustomizeURL =  "{{.StandardCRDsKustomizeURL}}"
-	GatewayCRDsKustomizeURL = "{{.CRDsKustomizeURL}}"
+	GatewayStandardCRDsKustomizeURL =  "{{.CRDsStandardKustomizeURL}}"
+	GatewayExperimentalCRDsKustomizeURL = "{{.CRDsExperimentalKustomizeURL}}"
 	GatewayRawRepoURL = "{{.RawRepoURL}}"
 )
